READ THIS
I created a FiveM script with two types of drug sales:

Through a dealer — the price is higher.

Directly to NPCs — the price is lower.

The script is optimized and uses 0.00ms CPU time. It also includes the /startquest command, where you can earn bonus money from drugs. The entire script is built on ox_lib and ox_target, except for /startquest, which is done using E.

If you find any bugs or problems, feel free to message me—I’m happy to help and fix any issues.

PS: This is my first script ever, so please be understanding!